First usable version of a color enhance filter
This commit creates a color enhancement GStreamer plugin. It is
basically a port of the Color Enhance 0.10 filter from The GIMP
2.6.11. The original code was converted to use integer calculations
only, for increased performance. The algorithm uses HSV color
space, so it needs to convert the RGB stream. I split the
conversion from the actual enhancement, and modified the algorithm
to do the conversion of the whole frame at once. A redundant
conversion pass could be removed this way. Real-time performance
is now reasonable on a Core 2 Duo T7100.