Create a despeckle filter
This commit creates a GStreamer despeckle filter. The core
algorithm is ported from the Despeckle filter of The GIMP. It is
basically an adjustable region size median filter that can
optionally adjust filtering region size based on the histogram
of the region. It can also optionally operate in a recursive mode,
in which filtered pixel data is also written back to the source
buffer, producing a different, and stronger, effect. Region size,
bright and dark thresholds, recursivity, and adaptivity can be
adjusted via properties.