Create a selective sharpening element
The element added in this commit takes an AYUV stream on one of its sink
pads, and a 8 bit grayscale video stream on the other. Global sharpening
is performed on the AYUV video stream, and the result is blended with the
original, using the grayscale mask as the blend ratio. This basically
means that the brighter a pixel is in the mask, the stronger the sharpening
effect is on the corresponding frame pixel. This is useful for
"smart sharpening", where the mask is an edge magnitude image stream.

The basic operation of the filter is the following.
The filter currently creates an internal bin with a gaussianblur element
and a tee. The sigma of the gaussianblur is set to a negative value for
sharpening. Three ghost pads are added to the bin: one sink for the original
frames, a source for getting the original frames back, and another source for
the sharpened frames. The frame source pad of the element itself is a ghost
pad targeted at the sink pad of the sharpening bin. The source pads of the
bin are linked to internal sink pads that are not added to anything, but
activated for push mode manually. (This looks like a dirty hack.) These sink
pads are added to a CollectPads, along with the mask sink pad of the filter,
for synchronization. The CollectPads callback function blends the original
and the sharpened frames using the mask frames as the blend ratio. The
resulting frames are pushed out on the source pad of the filter.