sobel: add absolute gradient magnitude, fix range conversion
An accurate method for calculating the gradient magnitude from x and y
gradient images is the G = sqrt(g_x*g_x + g_y*g_y) formula. There is a
little less accurate, but faster method called absolute gradient
magnitude: the formula G = abs(g_x) + abs(g_y). Support for the latter
was added through the abs-magnitude boolean property.

As the result from both formulas is well out of the range of uint8,
some kind of range conversion is needed to be able to store the results
in an output buffer. Scaling is the only option in current code. This
commit adds the option to clamp the value to [0,255] instead.

Also, scaling is implemented for absolute gradient magnitude, and fixed
for the original formula. Scaling ratios are currently empirically
determined values.

Finally, the empirically determined maxima are put in the .h file
as #defines, to reduce the amount of misleading "magic" numbers in
the code.