gimpdespeckle: canonicalize property names
There were underscores in property names in gimpdespeckle. As underscores
are not allowed in canonical names, these underscores were replaced by
dashes at runtime. This, after setting the G_PARAM_STATIC_STRINGS flag
on these properties, resulted in a segmentation fault when trying to
initialize the element's class. This commit replaces the underscores
with dashes in the code, and fixes the problem.