addalpha: handle EOS
A GstCollectPads returns NULL when popping a buffer from a pad on
which EOS was received. This leads to a segmentation fault in
current code, because this is not properly handled. This commit corrects
this issue by pushing a fully transparent, all zero frame on the source
pad of the addalpha element if the frame sink pad has received an EOS
event, pushing out the received original frame with all-one (fully opaque)
alpha channel if the mask sink pad has received an EOS, or pushing an EOS
event on the source pad if both sink pads have received an EOS event. The
latter is in accordance with the GstCollectPads documentation.